added features
Got the logic working for the water, just need to add in the 3D model and build upon the current code. (spawn rates, sugar crash)
Optimized the player movement code to suit good coding practices.
Almost done coding the flying monster, just need the 3d model from team and to modify the a part of the code for it to fly downwards.
Worked on the running/sliding/jumping animation as conan sent me a new fbx file with the new animations, cant get it to work correctly atm.